+++ crates/protocols/hpke/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.1] - 2026-03-20
+
+- Update crypto providers
+
 ## [0.6.0] - 2026-02-13
 
 ### Added